backend/cmda-hub-backend-legacy/src/main/resources/requirements.txt
Babel
Flask
matplotlib
nselib
nsepython
numpy
pandas
plotly
python_dateutil
scikit_learn
